base: ps2kbd: Ensure value remains unsigned when modifying CCB in i8042 probe
Fixes a compiler warning when compiling with GCC:
base/ps2kbd/ps2kbd.c:136:20: warning: unsigned conversion from 'int' to 'Byte' {aka 'unsigned char'} changes the value of '-82' [-Wsign-conversion]
  136 |         new_ccb &= ~((1 << 0) | (1 << 4) | (1 << 6));

+++ base/ps2kbd/ps2kbd.c
@@ -133,7 +133,7 @@ static Status Probe8042Controller(void) {
          * See https://wiki.osdev.org/I8042_PS/2_Controller#Initialising_the_PS/2_Controller to
          * understand the bit values here.
          */
-        new_ccb &= ~((1 << 0) | (1 << 4) | (1 << 6));
+        new_ccb &= (Byte)(~((1U << 0) | (1U << 4) | (1U << 6)));
         i8042Write(PS2_WRITE_CONFIG_BYTE);
         i8042WriteData(new_ccb);
